api: Move all new API logic to lit-signing-condition experiment
